+++ Globals/VerilogHelperFunctions.cs
@@ -1,0 +1,49 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+namespace VerilogLanguage
+{
+    public static partial class VerilogGlobals
+    {
+        // some characters, when encountered will need to have a full refresh, as they can have far-reaching consequences.
+        static public bool IsRefreshChar(string theString)
+        {
+            return (theString.Contains("/")) ||
+                   (theString.Contains("*")) ||
+                   (theString.Contains("[")) ||
+                   (theString.Contains("]")) ||
+                   (theString.Contains("}")) ||
+                   (theString.Contains("{")) ||
+                   (theString.Contains("(")) ||
+                   (theString.Contains(")"));
+        }
+
+        static public bool IsDelimeter(string theString)
+        {
+            return (theString == " ") ||
+                   (theString == "~") ||
+                   (theString == "[") ||
+                   (theString == "]") ||
+                   (theString == "}") ||
+                   (theString == "{") ||
+                   (theString == "(") ||
+                   (theString == ")") ||
+                   (theString == ";") ||
+                   (theString == ",") ||
+                   (theString == "@") ||
+                   (theString == "\"") ||
+                   (theString == "\t");
+        }
+
+        static public bool IsEndingDelimeter(string theString)
+        {
+            return (theString == "]") ||
+                   (theString == "}") ||
+                   (theString == ")");
+        }
+
+    }
+}
