+++ Globals/VerilogParseState.cs
@@ -1,0 +1,95 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+namespace VerilogLanguage
+{
+    public static partial class VerilogGlobals
+    {
+        /// <summary>
+        ///   VerilogParseState - while processing each segment, we'll keep track of attributes in a VerilogParseState 
+        /// </summary>
+        public struct VerilogParseState
+        {
+            public string thisItem;
+            public int thisIndex;
+            public string priorChar;
+            public string priorDelimiter;
+            public bool thisCharIsDelimiter;
+            public bool priorCharIsDelimiter;
+
+            public bool thisCharIsEndingDelimiter;
+            public bool priorCharIsIsEndingDelimiter;
+
+            public bool hasOpenSquareBracket;
+            public bool hasOpenRoundBracket;
+            public bool hasOpenSquigglyBracket;
+
+            public bool hasOpenDoubleQuote;
+
+            public bool IsNewDelimitedSegment;
+
+            private string _thisChar;
+            public string thisChar
+            {
+                get { return _thisChar; }
+                set
+                {
+                    _thisChar = value;
+                    thisCharIsDelimiter = IsDelimeter(value);
+                    thisCharIsEndingDelimiter = IsEndingDelimeter(value);
+                    priorCharIsDelimiter = IsDelimeter(priorChar);
+                    // note  contiguous spaces are a single segment
+                    IsNewDelimitedSegment = (thisCharIsDelimiter || priorCharIsDelimiter) && !((_thisChar == " ") && (priorChar == " "));
+
+                    if (IsNewDelimitedSegment)
+                    {
+
+                    }
+                    else
+                    {
+                        thisItem += thisChar;
+                    }
+                }
+            }
+
+            public void SetPriorValues()
+            {
+                // return;
+                priorCharIsDelimiter = thisCharIsDelimiter;
+                priorCharIsIsEndingDelimiter = thisCharIsEndingDelimiter;
+
+                priorChar = thisChar;
+                if (thisCharIsDelimiter)
+                {
+                    priorDelimiter = thisChar;
+                }
+            }
+
+
+            // initialize this VerilogParseState at creation time
+            public VerilogParseState(int i)
+            {
+                _thisChar = "";
+                thisIndex = 0;
+                thisItem = "";
+                IsNewDelimitedSegment = false;
+                priorChar = "";
+                priorDelimiter = "";
+                thisCharIsDelimiter = false;
+                priorCharIsDelimiter = false;
+                thisCharIsEndingDelimiter = false;
+                priorCharIsIsEndingDelimiter = false;
+                hasOpenSquareBracket = false;
+                hasOpenRoundBracket = false;
+                hasOpenSquigglyBracket = false;
+                hasOpenDoubleQuote = false;
+            }
+        }
+
+
+
+    }
+}
