+++ Globals/VerilogToken.cs
@@ -8,187 +8,6 @@ namespace VerilogLanguage
 {
     public static partial class VerilogGlobals
     {
-        public enum VerilogTokenContextType
-        {
-            Undetermined,
-            DoubleQuoteOpen,
-            SquareBracketOpen,
-            SquareBracketClose,
-            SquareBracketContents,
-            RoundBracketOpen,
-            RoundBracketClose,
-            RoundBracketContents,
-            SquigglyBracketOpen,
-            SquigglyBracketClose,
-            SquigglyBracketContents,
-            AlwaysAt,
-            Comment,
-            Text
-        }
-
-
-
-        // some characters, when encountered will need to have a full refresh, as they can have far-reaching consequences.
-        static public bool IsRefreshChar(string theString)
-        {
-            return (theString.Contains("/")) ||
-                   (theString.Contains("*")) ||
-                   (theString.Contains("[")) ||
-                   (theString.Contains("]")) ||
-                   (theString.Contains("}")) ||
-                   (theString.Contains("{")) ||
-                   (theString.Contains("(")) ||
-                   (theString.Contains(")"));
-        }
-
-        static public bool IsDelimeter(string theString)
-        {
-            return (theString == " ") ||
-                   (theString == "~") ||
-                   (theString == "[") ||
-                   (theString == "]") ||
-                   (theString == "}") ||
-                   (theString == "{") ||
-                   (theString == "(") ||
-                   (theString == ")") ||
-                   (theString == ";") ||
-                   (theString == ",") ||
-                   (theString == "@") ||
-                   (theString == "\"") ||
-                   (theString == "\t");
-        }
-
-        static public bool IsEndingDelimeter(string theString)
-        {
-            return (theString == "]") ||
-                   (theString == "}") ||
-                   (theString == ")");
-        }
-
-
-        /// <summary>
-        ///   VerilogParseState - while processing each segment, we'll keep track of attributes in a VerilogParseState 
-        /// </summary>
-        public struct VerilogParseState
-        {
-            public string thisItem;
-            public int thisIndex;
-            public string priorChar;
-            public string priorDelimiter;
-            public bool thisCharIsDelimiter;
-            public bool priorCharIsDelimiter;
-
-            public bool thisCharIsEndingDelimiter;
-            public bool priorCharIsIsEndingDelimiter;
-
-            public bool hasOpenSquareBracket;
-            public bool hasOpenRoundBracket;
-            public bool hasOpenSquigglyBracket;
-
-            public bool hasOpenDoubleQuote;
-
-            public bool IsNewDelimitedSegment;
-
-            private string _thisChar;
-            public string thisChar
-            {
-                get { return _thisChar; }
-                set
-                {
-                    _thisChar = value;
-                    thisCharIsDelimiter = IsDelimeter(value);
-                    thisCharIsEndingDelimiter = IsEndingDelimeter(value);
-                    priorCharIsDelimiter = IsDelimeter(priorChar);
-                    // note  contiguous spaces are a single segment
-                    IsNewDelimitedSegment = (thisCharIsDelimiter || priorCharIsDelimiter) && !((_thisChar == " ") && (priorChar == " "));
-
-                    if (IsNewDelimitedSegment)
-                    {
-
-                    }
-                    else
-                    {
-                        thisItem += thisChar;
-                    }
-                }
-            }
-
-            public void SetPriorValues()
-            {
-                // return;
-                priorCharIsDelimiter = thisCharIsDelimiter;
-                priorCharIsIsEndingDelimiter = thisCharIsEndingDelimiter;
-
-                priorChar = thisChar;
-                if (thisCharIsDelimiter)
-                {
-                    priorDelimiter = thisChar;
-                }
-            }
-
-
-            // initialize this VerilogParseState at creation time
-            public VerilogParseState(int i)
-            {
-                _thisChar = "";
-                thisIndex = 0;
-                thisItem = "";
-                IsNewDelimitedSegment = false;
-                priorChar = "";
-                priorDelimiter = "";
-                thisCharIsDelimiter = false;
-                priorCharIsDelimiter = false;
-                thisCharIsEndingDelimiter = false;
-                priorCharIsIsEndingDelimiter = false;
-                hasOpenSquareBracket = false;
-                hasOpenRoundBracket = false;
-                hasOpenSquigglyBracket = false;
-                hasOpenDoubleQuote = false;
-            }
-        }
-
-
-        private static VerilogTokenContextType VerilogTokenContextFromString(string s)
-        {
-            switch (s)
-            {
-                case null:
-                    return VerilogTokenContextType.Undetermined;
-
-                case "":
-                    return VerilogTokenContextType.Text;
-
-                default:
-                    switch (s.Substring(0, 1))
-                    {
-                        case "[":
-                            return VerilogTokenContextType.SquareBracketOpen;
-
-                        case "]":
-                            return VerilogTokenContextType.SquareBracketClose;
-
-                        case "(":
-                            return VerilogTokenContextType.RoundBracketOpen;
-
-                        case ")":
-                            return VerilogTokenContextType.RoundBracketClose;
-
-                        case "{":
-                            return VerilogTokenContextType.SquigglyBracketOpen;
-
-                        case "}":
-                            return VerilogTokenContextType.SquigglyBracketClose;
-
-                        case "@":
-                            return VerilogTokenContextType.AlwaysAt;
-
-                        default:
-                            return VerilogTokenContextType.Text;
-
-                    }
-            }
-        }
-
 
         /// <summary>
         ///   VerilogToken
